Refs #24330. simplified implementation per Andrei's request
